* Uses BasePlugin for registering plugins instead of Plugin (fix for apple leak BasePlugin provides better solution) * Moved SDK error types to PKError, now all error related information is in PKError. * Added @objc attribute to BasePlugin
